third_party/blink/web_tests/external/wpt/svg/text/scripted/getrotationofchar-expected.txt

This is a testharness.js-based test.
[FAIL] <textPath> - sideways-lr
  assert_equals: F expected 225 but got 45
Harness: the test ran to completion.
